Implementation of the print_nato(), send_morse() and send_code() functions in the main.py file. Also implementation of the main() function. The program is now fully functional and will produce the NATO Phonetic Alphabet for any given text as well as play the Morse code for that text.
symbols.py is updates with a complete directory for the symbols that will be used. It is easy to add additional symbols if needed. No changes need to be done in the main program to work with additional symbols. New symbols only needs to be added in the dictionary in symbols.py.

A helper program to create the dictionary for the symbols.py file was created. I already had the lists of the NATO Alphabet symbols and the Morse codes from a C program that I just made. Instead of typing everything in again I wrote the dict_creator.py program to create the dictionary for symbols.py. This helper program is also checked in.