Added Female Voice
It is now possible to choose between a male or a female voice for reading the NATO Phonetics of the text.

The male voice is using Windows speech engine. The female voice is using Google speech engine. Which voice to use is configured in the Speech Properties section at the top of the file.

Sections at the top of the file list configuration properties for printed text, Morse, and speech in different sections to make it easy to modify most behavior of the program without having to change any code.