Updated to use the new functions in nato.py
The commandline application is now using the two new functions that was created in nato.py for converting a text into NATO Phonetics and speaking the phonetics. The function convert_to_nato() does not have any side effects (print statements) and may be preferable to use in most cases, as it is done here.