Partially implementation of text conversion to NATO Phonetics.
Basic implementation of the conversion from text to NATO Phonetics. When clicking on the NATO button the text that is in the Input Text are is converted to NATO Phonetics and printed in the NATO Phonetics output area.

Only text is implemented. Voice is still missing. Also,all the switches and the language combobox remains to be implemented. But it is possible to test the text conversion to NATO Phonetics itself.