Changed the Appearance and Color Theme configuration selection to radio button menus.
Since you can only select one appearance or one color theme this should be implemented as a radiobutton menu, where a checkmark shows which option that is chosen. This is fixed now.

Also, a small update to the sliders behaviors where all changed to the sliders now are using the set functions at some point.