Running NATO Phonetics in its own thread.
When the GUI and the NATO Phonetics are running in the same thread the GUI will freeze during the time the text is converted to the phonetics and played in the speakers. If the text is long the GUI will be frozen for a long time. This is solved by letting the NATO Phonetics code run in its own thread. The functionality is there in this check-in. The Cancel button is not implemented.