Added CANCEL of NATO Phonetics.
If the voice is a male voice then the NATO button changes to a CANCEL button, and it is possible to cancel the speech before it is done.

If the voice is a female voice then the NATO button changes to DISABLED. Nothing happens when the button is pressed until sppech is done.

When the speech is done or has been cancelled the NATO button changes back to its normal state saing NATO again. A new conversion can be done.