First implementation of the Morse Button.
This is only a proof of concept to see that the button itself behaves properly and that Morse code is printed and played. At the moment this is visible in the console. Future work will show this in a separate window.

The basic behavior of the Morse Button works with this checkin. It runs in its own thread allowing other GUI functions to be performed, and it is possible to cancel the Morse presentation at any time.